Explicit vtordisp for MSVC
Summary:Because sometimes things are just different.

MSVC warns about the `dynamic_cast` done against this class because the class's declaration doesn't show that it would need a vtordisp. While I don't believe the `dynamic_cast` in question is actually called in either the constructor or destructor paths for the class, better to be safe and force the disp to be created.

Reviewed By: yfeldblum

Differential Revision: D2887490

Pulled By: Orvid

fb-gh-sync-id: 821a6fc820087ac7120fffd9f8b95a26c2b4e71b
shipit-source-id: 821a6fc820087ac7120fffd9f8b95a26c2b4e71b