folly: fix gflags<->glog configure check ordering
Summary:On systems that do not have gflags or glog installed globally,
but only locally as static libraries, our configure check failed for
glog because we were checking it prior to gflags.

glog depends on gflags, but when the libraries are static, there is
no explicit dependency information available and the linker fails.

This resolves the issue by checking for gflags first; this causes
configure to add an implicit `-lgflags` for the subsequent glog
tests.

Reviewed By: yfeldblum

Differential Revision: D3044138

fb-gh-sync-id: 5e07ce52842c6e4cff796560672bf950e2fafe6c
shipit-source-id: 5e07ce52842c6e4cff796560672bf950e2fafe6c